master states/templates: 237 nginx now starts with supervisor

+++ setup/templates/proxy/conf/salt-startup.sh
@@ -1,6 +1,9 @@
 #!/bin/bash
 
-# we need to remove the master pub key becase the
-# startup of this machine envokes new not accepted
+# we need to remove the master pub key because the
+# startup of this machine invokes new not accepted
 # crypto stuff
-rm /var/lib/salt/pki/minion/minion_master.pub
+if [ -f "/var/lib/salt/pki/minion/minion_master.pub" ]
+then
+  rm /var/lib/salt/pki/minion/minion_master.pub
+fi
